got the connection and some of the promting done
